Extract applet-ai core
